[avalanche] Add regression testing for flip flopping votes on blocks
Summary:
As we introduce new behavior (block policies) that could impact block parking
we need to ensure that flip flopping avalanche votes behave as expected and do
not regress.

Depends on D13149

I also took this opportunity to make the log checking more consistent for the
first acceptance.

Test Plan:
  ./test/functional/test_runner.py abc_p2p_avalanche_voting

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13150