remove most of versionbits / BIP9 related code
Summary:
The only remaining piece is `ComputeBlockVersion` and code related to `BuriedDeployments`

The functional test mining_basic.py needs a change because the removed TESTDUMMY deployment was always active on testnet, and this is the only place where it was tested. There is no need to keep this test.

Depends on D13157

Test Plan: `ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13158