[avalanche] Log when a proof becomes invalid
Summary:
This is useful in a lot of cases, especially when a proof expires.

Especially acute in the test environment where many proofs may have the same
expiry for convenience. I learned the hard way that proofs unexpectedly
expiring leads to no other clues in the logs other than simply not polling.

[../src/avalanche/peermanager.cpp:501] [updatedBlockTip] Peer proof 37011d056427f4639c80dc3df324fcfe26d98f8ffb6b12a016d52a6d37e9d0ef verification failed (expired-proof)

Test Plan:
  ninja && ./test/functional/test_runner.py --nocleanup abc_rpc_getavalancheinfo

Log appears like this:
  [../src/avalanche/peermanager.cpp:500] [updatedBlockTip] Invalidating proof 70d67e4e03555ade4f40944236703e076e04b8ddb4fb5c9c10cc871667e7efa8: verification failed (expired-proof)

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13160