[avalanche] Disable test framework default proof expiry
Summary:
For picking test timestamps "far in the future", it is natural to grab the
current timestamp, increment the most signficant digit and zero out the rest.
Until 2033 that process results in 2000000000.

The current test framework default proof expiry just happens to be this time
and so do various activation times that we tend to use. This is bad because
it leads to unnecessary debugging that could have been avoided with a good
default. And if you are really unlucky, other issues make that debugging
experience much worse: D13160

abc_rpc_getavalancheinfo.py has been modified because it depends on some
non-zero expiry for the test.

Test Plan:
  ninja check-functional

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13161