Latch the wellington activation in the mempool
Summary:
After wellington we will stop computing the ancestors and descendant
limits, and stom enforcing them. Because the mempool has no idea what the
current tip is (and it should not), we will use an activation latch so it
knows when the statistics can be skipped. This also makes an hypothetical
reorg event easier to handle.

Inspired by [[https://gitlab.com/bitcoin-cash-node/bitcoin-cash-node/merge_requests/1117 | bchn#1117]].

Test Plan:
  ninja all check

Reviewers: #bitcoin_abc, tobias_ruck, PiRK, sdulfari

Reviewed By: #bitcoin_abc, tobias_ruck, PiRK, sdulfari

Subscribers: tobias_ruck

Differential Revision: https://reviews.bitcoinabc.org/D13170