Fix wait_for_debug_log acting as a noop
Summary:
wait_for_debug_log is not called using a with statement so the yield returns
the rest of the function as a generator that is never run.

Test Plan:
Wrap the `"Avalanche invalidated block {tip_to_park}"` check with:
```
with node.assert_debug_log([f"Avalanche invalidated block {tip_to_park}"]):
```

This fails!
  ninja && ./test/functional/test_runner.py abc_p2p_avalanche_voting

Apply this patch and see that it now succeeds as expected.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: PiRK, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13172