skip subtest test_work_queue_exceeded in interface_rpc.py when CLI is not compiled
Summary:
This test relies on sending an excessive number of RPC calls via the command line (`node.cli('getrpcinfo').send_cli())`). I tried sending the RPC calls with `node.getrpcinfo()` but that does not seem to work, the test keeps running forever.

I could not find a rationale in the original PR for why the CLI is used, but my testing seems to indicate that there is a good reason.

This unbreaks the `build-without-cli` CI build.

Test Plan:
`ninja check-functional`

`@bot build-without-cli`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13205