[Cashtab] [Alias] Real time update of Registered Aliases list upon new registration
Summary: Currently when registering a new alias it won't appear in the `Registered Aliases` section of Alias.js until 1) alias cache has re-synchronized and 2) page refresh. This diff ensures both occurs upon successful broadcast of a registration so users are not left wondering whether their registration was successful or not.

Test Plan:
- enable Alias to true in ticker.js
- npm start
- navigate to Alias.js and register a new alias. Observe it is displayed

Reviewers: bytesofman, #bitcoin_abc

Reviewed By: bytesofman, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D13264