test: Use self.wait_until over wait_until_helper
Summary:
`wait_until_helper` is a "private" helper, not intended to be used directly, because it doesn't scale the timeout with the timeout factor. Fix this by replacing it with a call to `self.wait_until`, which does the scaling.

Backport of [[https://github.com/bitcoin/bitcoin/pull/27226 | core#27226]]

Test Plan:
  ninja check-functional

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13527