[tg bot] Add test script for sending messages
Summary:
T2972

Depends on D13279

It's not possible to tell whether or not a given markdown string will send in Telegram without throwing it at the API. This requires an API call and manual review, so should not be integrated into CI.

Instead, added new script, which can be run with `npm run msgs`. This way the dev can live test usign the Telegram API and a dev bot in a dev channel.

Test Plan:
`npm run test`
`npm run msgs` and messages all send

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D13280