[alias-server] Cache confirmed tx history of registration address
Summary:
T2966

Depends on D13314

Store all tx history of the alias registration address in the database. When a new block is found -- if there are new confirmed txs at the alias registration address that are not in the database, add them.

Test Plan:
Review new mocks and tests
`npm test`
`node index.js` and all tx history should be added to db
Send some txs to the alias address
When they are confirmed, logs should confirm they are added to the db

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D13316