[guix] implement guix build for linux
Summary:
This is the first step of importing the guix build machinery from Bitcoin Core and make it work for Bitcoin ABC with its additonal dependencies.

This currently supports only HOST x86_64-linux-gnu. It does not include the guix-codesign, guix-verify and guix-attest scripts.
The guix-clean script has been simplified significantly.
The build is currently not deterministic for an unknown reason (tested on the same computer twice in a row).

Test Plan: Follow the steps of the "Guix Installation and Setup" doc, then run `contrib/guix/guix-build`, then test the binaries (bitcoind, bitcoin-cli) locally.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13764