Fix the makeseeds.py version filter
Summary: This was still accepting 0.25.x seeds depite they're no longer able to sync. Also 0.27.0 is out now, so let's add it to the filter.

Test Plan: There is no easy way to test without a complete seeder setup, but this is trivial and the result visible upon next automated seed update.

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D13393