[cmake] Don't select an alternative linker for the native build
Summary:
This non-default linker selection (either gold or lld) is not enforced during cross compilation, and can be easily overriden for standard build.
However there is an exception to this rule: during cross-compilation, the automatic selection applies for the native (host) builds. If the host has no gold/lld installed this could cause the native build to fail.

Test Plan:
  ninja
Run a cross build.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D13451