[ecashaddrjs] Version bump
Summary:
T3039

Depends on D13478

Bump minor version number as new functionality (support for decoding prefixless but correctly summed addresses) has been added in a backwards compatible manner.

Test Plan:
When this diff is green, I will publish this package as is in this diff to `npm`. It's good practice to make sure the package itself actually works. One approach here would be

```
cd web/cashtab
npm i ../ecashaddrjs
npm start
```
Confirm that Cashtab starts up and works as expected. Enter a valid cashaddress in the "to" field on the Send screen and confirm it is accepted. Change a character and confirm it is recognized as invalid.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13479