[Chronik] Fix: `tx_writer.insert` in `handle_block_disconnected`
Summary:
`chronik_tx` didn't include a test for when a block gets removed. This allowed an undetected oversight to slip in: When disconnecting a block, instead of removing txs from the tx index, we would add them.

This fixes it and adds a test to handle this case.

Test Plan: `ninja && ./test/functional/test_runner.py chronik_tx`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D13488