[Chronik] Move `mempool` module to `chronik-db`
Summary:
The mempool could be seen as a kind-of in-memory DB for unconfirmed txs, so this certainly isn't out of place.

Having the mempool here also allows for better synergies, e.g. reusing functions specific to `Group`s.

The legacy Chronik implementation also has the mempool in the `chronik-db` crate, somehow this was overlooked when porting.

Depends on D13487.

Test Plan: `ninja check-crates`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D13500