[Chronik] Add `MempoolGroupHistory`
Summary:
The equivalent of `GroupHistoryWriter/-Reader` but for the mempool.

We store the transactions by serialized group member. Txs within a member are sorted by time_first_seen first and then by txid, which makes it very easy to read the data in chronological order.

We don't employ any pagination because it wouldn't add any benefit (the data is already in RAM), and BTreeSet itself is a hierarchical structure already.

Depends on D13500.

Test Plan: `ninja check-crates`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D13501