[Chronik] Move `HashedError` to `error` module, rename to `DataError`
Summary:
When we add more primitives (e.g. PubKey), it's better to share one error type.

We don't use Report in bitcoinsuite-core, because we eventually want others to be able to use this crate, and we shouldn't force them to use Report for their error handling.

Test Plan: `ninja check-crates`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D13507