[Cashtab] Deprecate husky
Summary:
T3068

Husky is interfering with other javascript repos in the monorepo (annoying to dev on more than one as it tries to run changed cashtab unit tests, of which there are none).

The implementation was already a bit forced due to monorepo. Husky is not the right tool here.

The snapshot situation is annoying but it is caught by CI and husky does not prevent snapshot errors 100% of the time. Need a better approach.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D13514