[alias-server] Rename secrets.js.sample so that it can be linted and formatted
Summary:
T3060

This is a template file. It is supposed to have the same style and format as `secrets.js`, to be created by a running developer.

Keeping it as a non-js file means no one can use it accidentally. But this is already handled by fake (yet readable) data in the ID fields.

Rename to *.js so it is formatted appropriately. Add 'use strict' to pass linter.

Test Plan: Confirm this is just a name change of an unused template file.

Reviewers: #bitcoin_abc, emack

Reviewed By: #bitcoin_abc, emack

Differential Revision: https://reviews.bitcoinabc.org/D13515