[Chronik] Add `/script/:type/:payload/unconfirmed-txs`
Summary:
This is the equivalent of `/script/:type/:payload/confirmed-txs` but only for mempool txs. We (currently) always return all txs of a script without pagination, as querying from the mempool (which is in RAM) will always be fast.

We still return `TxHistoryPage`, in case we might need pagination in the future, and having it return the same type as the other endpoints is nice for users of the API.

Depends on D13567.

Test Plan: `ninja check-crates && ninja check-functional`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Differential Revision: https://reviews.bitcoinabc.org/D13574