[monorepo organization] rename xecjs-message to ecashjs-message
Summary:
T3067

Rename `xecjs-message` to `ecashjs-message`. This follows the standard of replacing the name `bitcoin` in the repo this is forked from with the long-form version of this cryptocurrency rather than the ticker.

Test Plan:
Confirm directory name and `package.json` names are changed.
Note: the README still says `bitcoinjs-message` and has lots of other legacy problems. This is to be handled in a separate diff.
Note: the unit tests don't work before this rename. This is to be handled in a separate diff.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13571