[Chronik] Add general-purpose python `ChronikClient`
Summary:
A simple client for querying the different Chronik endpoints in a nice flowing interface.

While it's currently placed in test_framework, in the future it is intended to be published on pip so people who want to access Chronik can use this client. For this reason we place it in its own directory.

We also update all the chronik tests to use this new client now.

Test Plan: `ninja check-functional`

Reviewers: Fabien, #bitcoin_abc

Reviewed By: Fabien, #bitcoin_abc

Subscribers: aimal

Differential Revision: https://reviews.bitcoinabc.org/D13612