[herald] Unit tests for parseWebsocketMessage
Summary:
T2972

Depends on D13616

Add unit tests for `parseWebsocketMessage`

These unit tests required a mock Telegram bot and improvements to `chronikMock.js`

For now, I think it's okay to maintain `chronikMock.js` as distinct files in both `alias-server` and `ecash-herald.` The incremental work in both apps is a good way to learn what features this mock needs. It will eventually be extended to Cashtab and live as a centralized dependency.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13617