[tg bot] Improve how mocks are generated and stored
Summary:
T2972

Depends on D13623

It's been a persistent headache that `generateMocks` uses slightly different syntax / functions than what is used in the app. Some differences are necessary to preserve blockNames, which are useful in developer review and keeping track of why some blocks are mocked. But, in general, we want to be sure that the mocks are generated in exactly the same way that the app is running.

This diff refactors `generateMocks` to use the new `handleBlockConnected` function, modified with an optional flag to return mocks.

Test Plan: `npm test`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D13624