[CI] Add support for running website containers
Summary:
This is a first step toward building on-demand temporarly website hosting as a preview. A config for e.cash is added to demonstrate the feature.

Note 1: you need to install the `at` utility for the preview feature to work.
Note 2: For now there is no port mangling, which means that you have to wait for the preview to expire before starting another one.

Test Plan:
  ./contrib/teamcity/build-configurations.py preview-e.cash
Navigate to localhost:3000 and check the website renders.

Reviewers: #bitcoin_abc, sdulfari

Reviewed By: #bitcoin_abc, sdulfari

Subscribers: sdulfari

Differential Revision: https://reviews.bitcoinabc.org/D13653