Bump minimum supported mypy version to 0.910
Summary: D13649 Introduced --install-types (mypy 0.810) and --non-interactive (mypy 0.910) so lets bump the version.

Test Plan:
  arc lint test/functional/abc_rpc_isfinal
Works on Debian 11.
Installed successfully on Debian 10, but I'm getting some lint errors.
No version of mypy up to 1.2 appeared to solve the issue so it is unclear
if it is my environment or something else. If someone could test on
another Debian 10 box for sanity that would be nice.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D13659